feat: explore state management (#4171)
* initial

* add jotai

* refile

* save file

* change style

* reset filter string query

* Update settings.json

* location
