+++ src/components/Explore/TokenDetails/Resource.tsx
