+++ src/components/Explore/TokenTable/TimeSelector.tsx
@@ -1,26 +1,33 @@
-import { getChainInfo } from 'constants/chainInfo'
-import { SupportedChainId } from 'constants/chains'
 import { useOnClickOutside } from 'hooks/useOnClickOutside'
-import { useRef, useState } from 'react'
+import { TimePeriod } from 'hooks/useTopTokens'
+import { useAtom } from 'jotai'
+import { useRef } from 'react'
 import { Check, ChevronDown, ChevronUp } from 'react-feather'
 import { useModalIsOpen, useToggleModal } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
-import styled, { useTheme } from 'styled-components/macro'
+import styled, { css, useTheme } from 'styled-components/macro'
 
-import { MEDIUM_MEDIA_BREAKPOINT } from './constants'
+import { filterTimeAtom } from '../state'
 
-const NETWORKS = [
-  SupportedChainId.MAINNET,
-  SupportedChainId.ARBITRUM_ONE,
-  SupportedChainId.POLYGON,
-  SupportedChainId.OPTIMISM,
-]
+export const TIME_DISPLAYS: { [key: string]: string } = {
+  hour: '1H',
+  day: '1D',
+  week: '1W',
+  month: '1M',
+  year: '1Y',
+}
+
+const TIMES = Object.values(TimePeriod)
+
+enum FlyoutAlignment {
+  LEFT = 'LEFT',
+  RIGHT = 'RIGHT',
+}
 
 const InternalMenuItem = styled.div`
   flex: 1;
-  padding: 12px 16px;
+  padding: 8px;
   color: ${({ theme }) => theme.textPrimary};
-
   :hover {
     cursor: pointer;
     text-decoration: none;
@@ -32,7 +39,9 @@ const InternalMenuItem = styled.div`
 
 const InternalLinkMenuItem = styled(InternalMenuItem)`
   display: flex;
+  flex-direction: row;
   align-items: center;
+  padding: 12px 16px;
   justify-content: space-between;
   text-decoration: none;
   cursor: pointer;
@@ -42,15 +51,15 @@ const InternalLinkMenuItem = styled(InternalMenuItem)`
     text-decoration: none;
   }
 `
-const MenuTimeFlyout = styled.span`
-  min-width: 200px;
+const MenuTimeFlyout = styled.span<{ flyoutAlignment?: FlyoutAlignment }>`
+  min-width: 150px;
   max-height: 350px;
   overflow: auto;
   background-color: ${({ theme }) => theme.backgroundSurface};
   box-shadow: 0px 0px 1px rgba(0, 0, 0, 0.01), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04),
     0px 24px 32px rgba(0, 0, 0, 0.01);
   border: 1px solid ${({ theme }) => theme.backgroundOutline};
-  border-radius: 12px;
+  border-radius: 16px;
   padding: 8px 0px;
   display: flex;
   flex-direction: column;
@@ -58,8 +67,15 @@ const MenuTimeFlyout = styled.span`
   position: absolute;
   top: 48px;
   z-index: 100;
-  left: 0px;
 
+  ${({ flyoutAlignment = FlyoutAlignment.RIGHT }) =>
+    flyoutAlignment === FlyoutAlignment.RIGHT
+      ? css`
+          right: 0px;
+        `
+      : css`
+          left: 0px;
+        `};
   ${({ theme }) => theme.mediaWidth.upToMedium`
     bottom: unset;
     right: 0;
@@ -70,19 +86,19 @@ const MenuTimeFlyout = styled.span`
 const StyledMenuButton = styled.button<{ open: boolean }>`
   width: 100%;
   height: 100%;
-  color: ${({ theme, open }) => (open ? theme.blue200 : theme.textPrimary)};
+  border: none;
   background-color: transparent;
+  color: ${({ theme, open }) => (open ? theme.blue200 : theme.textPrimary)};
+  margin: 0;
   background-color: ${({ theme, open }) => (open ? theme.accentActionSoft : theme.none)};
   border: 1px solid ${({ theme, open }) => (open ? theme.accentActiveSoft : theme.backgroundOutline)};
-  margin: 0;
   padding: 6px 12px 6px 12px;
   border-radius: 12px;
   font-size: 16px;
   line-height: 24px;
-  font-weight: 400;
+  font-weight: 600;
 
-  :hover,
-  :focus {
+  :hover {
     cursor: pointer;
     outline: none;
     background-color: ${({ theme, open }) => !open && theme.backgroundContainer};
@@ -100,11 +116,7 @@ const StyledMenu = styled.div`
   position: relative;
   border: none;
   text-align: left;
-  width: 160px;
-
-  @media only screen and (max-width: ${MEDIUM_MEDIA_BREAKPOINT}) {
-    flex: 1;
-  }
+  width: 80px;
 `
 
 const StyledMenuContent = styled.div`
@@ -113,7 +125,6 @@ const StyledMenuContent = styled.div`
   align-items: center;
   border: none;
   width: 100%;
-  font-weight: 600;
   vertical-align: middle;
 `
 
@@ -121,60 +132,39 @@ const Chevron = styled.span<{ open: boolean }>`
   padding-top: 1px;
   color: ${({ open, theme }) => (open ? theme.blue200 : theme.textPrimary)};
 `
-const NetworkLabel = styled.div`
-  display: flex;
-  gap: 8px;
-  align-items: center;
-`
-const Logo = styled.img`
-  height: 20px;
-  width: 20px;
-`
-const CheckContainer = styled.div`
-  display: flex;
-  flex-direction: flex-end;
-`
 
 // TODO: change this to reflect data pipeline
-export default function NetworkFilter() {
+export default function TimeSelector() {
   const theme = useTheme()
   const node = useRef<HTMLDivElement | null>(null)
-  const open = useModalIsOpen(ApplicationModal.NETWORK_FILTER)
-  const toggleMenu = useToggleModal(ApplicationModal.NETWORK_FILTER)
+  const open = useModalIsOpen(ApplicationModal.TIME_SELECTOR)
+  const toggleMenu = useToggleModal(ApplicationModal.TIME_SELECTOR)
   useOnClickOutside(node, open ? toggleMenu : undefined)
-  const [activeNetwork, setNetwork] = useState(SupportedChainId.MAINNET)
-  const { label, logoUrl } = getChainInfo(activeNetwork)
+  const [activeTime, setTime] = useAtom(filterTimeAtom)
 
   return (
     <StyledMenu ref={node}>
-      <StyledMenuButton onClick={toggleMenu} aria-label={`networkFilter`} open={open}>
+      <StyledMenuButton onClick={toggleMenu} aria-label={`timeSelector`} open={open}>
         <StyledMenuContent>
-          <NetworkLabel>
-            <Logo src={logoUrl} /> {label}
-          </NetworkLabel>
+          {TIME_DISPLAYS[activeTime]}
           <Chevron open={open}>
             {open ? <ChevronUp size={15} viewBox="0 0 24 20" /> : <ChevronDown size={15} viewBox="0 0 24 20" />}
           </Chevron>
         </StyledMenuContent>
       </StyledMenuButton>
+      {/* handles the actual flyout of the menu*/}
       {open && (
         <MenuTimeFlyout>
-          {NETWORKS.map((network) => (
+          {TIMES.map((time) => (
             <InternalLinkMenuItem
-              key={network}
+              key={time}
               onClick={() => {
-                setNetwork(network)
+                setTime(time)
                 toggleMenu()
               }}
             >
-              <NetworkLabel>
-                <Logo src={getChainInfo(network).logoUrl} /> {getChainInfo(network).label}
-              </NetworkLabel>
-              {network === activeNetwork && (
-                <CheckContainer>
-                  <Check size={16} color={theme.accentAction} />
-                </CheckContainer>
-              )}
+              <div>{TIME_DISPLAYS[time]}</div>
+              {time === activeTime && <Check color={theme.accentAction} size={16} />}
             </InternalLinkMenuItem>
           ))}
         </MenuTimeFlyout>
