+++ src/components/Explore/TokenTable/TokenRow.tsx
@@ -4,6 +4,7 @@ import { useCurrency, useToken } from 'hooks/Tokens'
 import useTheme from 'hooks/useTheme'
 import { TimePeriod, TokenData } from 'hooks/useTopTokens'
 import { useAtom } from 'jotai'
+import { useAtomValue } from 'jotai/utils'
 import { ReactNode } from 'react'
 import { ArrowDown, ArrowDownRight, ArrowUp, ArrowUpRight, Heart } from 'react-feather'
 import { Link } from 'react-router-dom'
@@ -16,21 +17,17 @@ import {
   MEDIUM_MEDIA_BREAKPOINT,
   MOBILE_MEDIA_BREAKPOINT,
   SMALL_MEDIA_BREAKPOINT,
-} from './constants'
-import { favoritesAtom, useToggleFavorite } from './state'
+} from '../constants'
+import {
+  favoritesAtom,
+  filterTimeAtom,
+  sortCategoryAtom,
+  sortDirectionAtom,
+  useSetSortCategory,
+  useToggleFavorite,
+} from '../state'
+import { Category, SortDirection } from '../types'
 import { TIME_DISPLAYS } from './TimeSelector'
-
-enum Category {
-  percent_change = '% Change',
-  market_cap = 'Market Cap',
-  price = 'Price',
-  volume = 'Volume',
-}
-enum SortDirection {
-  Increasing = 'Increasing',
-  Decreasing = 'Decreasing',
-}
-const SORT_CATEGORIES = Object.values(Category)
 
 const ArrowCell = styled.div`
   padding-left: 2px;
@@ -80,11 +77,11 @@ const StyledTokenRow = styled.div`
   }
 
   @media only screen and (max-width: ${MOBILE_MEDIA_BREAKPOINT}) {
-    grid-template-columns: 1fr 12fr 6fr;
+    grid-template-columns: 12fr 6fr;
     width: fit-content;
     min-width: unset;
     border-bottom: 0.5px solid ${({ theme }) => theme.backgroundContainer};
-    padding: 0px;
+    padding: 0px 12px;
 
     :last-of-type {
       border-bottom: none;
@@ -143,9 +140,7 @@ const ListNumberCell = styled(Cell)`
   min-width: 32px;
 
   @media only screen and (max-width: ${MOBILE_MEDIA_BREAKPOINT}) {
-    font-size: 12px;
-    justify-content: flex-start;
-    min-width: 20px;
+    display: none;
   }
 `
 const MarketCapCell = styled(Cell)<{ sortable: boolean }>`
@@ -318,23 +313,22 @@ function getHeaderDisplay(category: string, timeframe: string): string {
 /* Get singular header cell for header row */
 function HeaderCell({
   category,
-  sortDirection,
-  isSorted,
   sortable,
-  timeframe,
 }: {
-  category: string // TODO: change this to make it work for trans
-  sortDirection: SortDirection
-  isSorted: boolean
+  category: Category // TODO: change this to make it work for trans
   sortable: boolean
-  timeframe: string
 }) {
   const theme = useTheme()
-  if (isSorted) {
+  const sortDirection = useAtomValue<SortDirection>(sortDirectionAtom)
+  const handleSortCategory = useSetSortCategory(category)
+  const sortCategory = useAtomValue<Category>(sortCategoryAtom)
+  const timeframe = useAtomValue<TimePeriod>(filterTimeAtom)
+
+  if (sortCategory === category) {
     return (
-      <SortingCategory>
+      <SortingCategory onClick={handleSortCategory}>
         <SortArrowCell>
-          {sortDirection === SortDirection.Decreasing ? (
+          {sortDirection === SortDirection.increasing ? (
             <ArrowDown size={14} color={theme.accentActive} />
           ) : (
             <ArrowUp size={14} color={theme.accentActive} />
@@ -344,7 +338,7 @@ function HeaderCell({
       </SortingCategory>
     )
   }
-  if (sortable) return <SortOption>{getHeaderDisplay(category, timeframe)}</SortOption>
+  if (sortable) return <SortOption onClick={handleSortCategory}>{getHeaderDisplay(category, timeframe)}</SortOption>
   return <Trans>{getHeaderDisplay(category, timeframe)}</Trans>
 }
 
@@ -389,9 +383,9 @@ export function TokenRow({
 }
 
 /* Header Row: top header row component for table */
-export function HeaderRow({ timeframe }: { timeframe: string }) {
+export function HeaderRow() {
   /* TODO: access which sort category used and timeframe used (temporarily hardcoded values) */
-  const sortedBy = SORT_CATEGORIES[1]
+
   return (
     <TokenRow
       address={null}
@@ -399,42 +393,10 @@ export function HeaderRow({ timeframe }: { timeframe: string }) {
       favorited={null}
       listNumber={null}
       tokenInfo={<Trans>Name</Trans>}
-      price={
-        <HeaderCell
-          category={Category.price}
-          sortDirection={SortDirection.Decreasing}
-          isSorted={sortedBy === Category.price}
-          sortable
-          timeframe={timeframe}
-        />
-      }
-      percentChange={
-        <HeaderCell
-          category={Category.percent_change}
-          sortDirection={SortDirection.Decreasing}
-          isSorted={sortedBy === Category.percent_change}
-          sortable
-          timeframe={timeframe}
-        />
-      }
-      marketCap={
-        <HeaderCell
-          category={Category.market_cap}
-          sortDirection={SortDirection.Decreasing}
-          isSorted={sortedBy === Category.market_cap}
-          sortable
-          timeframe={timeframe}
-        />
-      }
-      volume={
-        <HeaderCell
-          category={Category.volume}
-          sortDirection={SortDirection.Decreasing}
-          isSorted={sortedBy === Category.volume}
-          sortable
-          timeframe={timeframe}
-        />
-      }
+      price={<HeaderCell category={Category.price} sortable />}
+      percentChange={<HeaderCell category={Category.percentChange} sortable />}
+      marketCap={<HeaderCell category={Category.marketCap} sortable />}
+      volume={<HeaderCell category={Category.volume} sortable />}
       sparkLine={null}
     />
   )
